CSV_Analyzer_AI/requirements.txt
streamlit
pandas
langchain
langgraph
langchain-google-genai
matplotlib
openai
python-dotenv
